Move from backpack to data families and nullary type classes
