+++ wallet/ergvein-wallet.cabal
@@ -14,48 +14,32 @@ flag android
   default: False
   manual: True
 
-library ergvein-wallet-language
-  hs-source-dirs:      src
+library ergvein-wallet-native
+  hs-source-dirs:      src-native
   exposed-modules:
-    Reflex.Localize.Language
-  build-depends:
-      base
-  default-language:    Haskell2010
-  default-extensions:
-    DeriveGeneric
-
-library ergvein-wallet-callbacks
-  hs-source-dirs:      src-callbacks
-  exposed-modules:
+    Ergvein.Wallet.Native
+    Ergvein.Wallet.Run
     Ergvein.Wallet.Run.Callbacks
   default-language:    Haskell2010
   build-depends:
       base
-
-library ergvein-wallet-sigs
-  hs-source-dirs:      src-hsig
-  signatures:
-    Ergvein.Wallet.Run
-    Ergvein.Wallet.Native
-  default-language:    Haskell2010
-  build-depends:
-      base
     , aeson
-    , ergvein-wallet-callbacks
     , jsaddle
     , text
+  default-extensions:
+    MultiParamTypeClasses
 
 library ergvein-wallet-android
   hs-source-dirs:      src-android
   exposed-modules:
-    Ergvein.Wallet.Native
-    Ergvein.Wallet.Native.Android.MainWidget
-    Ergvein.Wallet.Run
+    Ergvein.Wallet.Android.Native
+    Ergvein.Wallet.Android.Native.MainWidget
+    Ergvein.Wallet.Android.Run
   build-depends:
       base
     , jsaddle
     , android-activity
-    , ergvein-wallet-callbacks
+    , ergvein-wallet-native
   default-language:    Haskell2010
   c-sources:
     cbits/android.c
@@ -71,16 +55,18 @@ library ergvein-wallet-android
 library ergvein-wallet-desktop
   hs-source-dirs:      src-desktop
   exposed-modules:
-    Ergvein.Wallet.Native
-    Ergvein.Wallet.Run
+    Ergvein.Wallet.Desktop.Native
+    Ergvein.Wallet.Desktop.Run
   build-depends:
       base
     , aeson
-    , ergvein-wallet-callbacks
+    , ergvein-wallet-native
     , jsaddle
     , jsaddle-webkit2gtk
-    , text 
+    , text
   default-language:    Haskell2010
+  default-extensions:
+    MultiParamTypeClasses
 
 library
   hs-source-dirs:      src
@@ -91,6 +77,7 @@ library
     Ergvein.Wallet.Embed.TH
     Ergvein.Wallet.Id
     Ergvein.Wallet.Input
+    Ergvein.Wallet.Language
     Ergvein.Wallet.Main
     Ergvein.Wallet.Monad
     Ergvein.Wallet.Monad.Env
@@ -116,9 +103,7 @@ library
     , directory
     , ergvein-common            >= 1.0      && < 1.1
     , ergvein-crypto            >= 0.1      && < 0.2
-    , ergvein-wallet-callbacks
-    , ergvein-wallet-language
-    , ergvein-wallet-sigs
+    , ergvein-wallet-native
     , exceptions
     , file-embed                >= 0.0      && < 0.1
     , ghcjs-dom
@@ -176,6 +161,7 @@ executable ergvein
       base
     , data-default
     , ergvein-wallet
+    , ergvein-wallet-native
     , optparse-generic  >= 1.3      && < 1.4
     , reflex
     , reflex-dom
