+++ wallet/src-android/Ergvein/Wallet/Android/Native.hs
@@ -1,0 +1,3 @@
+module Ergvein.Wallet.Native.Android(
+
+  ) where
