+++ wallet/src-android/Ergvein/Wallet/Android/Native/MainWidget.hsc
