+++ wallet/src-native/Ergvein/Wallet/Run/Callbacks.hs
