+++ wallet/src/Ergvein/Wallet/Language.hs
@@ -1,12 +1,13 @@
--- | Module that fullfils signature required by `reflex-localize`
-module Reflex.Localize.Language(
+-- | Module that fullfils data family required by `reflex-localize`
+module Ergvein.Wallet.Language(
     Language(..)
   ) where
 
 import GHC.Generics (Generic)
+import Reflex.Localize.Language
 
 -- | Languages that are supported by wallet
-data Language
+data instance Language
   = English
   | Russian
   deriving (Eq, Ord, Show, Read, Generic)
